% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdist.R
\name{rdist}
\alias{rdist}
\title{sumulation d'une loi discrete}
\usage{
rdist(x, p)
}
\arguments{
\item{x}{numeric vector}

\item{p}{numeric vector}
}
\description{
sumulation d'une loi discrete
}
